pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-outreach-automation"
version = "0.1.0"
description = "Professional LinkedIn Outreach Automation Platform"
readme = "README.md"
authors = [{name = "Lead Gen Professional", email = "contact@leadgen.automation"}]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "selenium>=4.15.2",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "click>=8.1.7",
    "pydantic>=2.5.0",
    "faker>=20.1.0",
    "webdriver-manager>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0", 
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
linkedin-automation = "src.cli.main:cli"

[project.urls]
Homepage = "https://github.com/leadgen/linkedin-outreach-automation"
Repository = "https://github.com/leadgen/linkedin-outreach-automation.git"
Issues = "https://github.com/leadgen/linkedin-outreach-automation/issues"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]